+++ src/nodes/epsrHelpers.cpp
@@ -10,7 +10,7 @@
 #include "nodes/dissolve.h"
 #include "nodes/epsr.h"
 #include "nodes/neutronSQ.h"
-#include "nodes/sq/sq.h"
+#include "nodes/sq.h"
 #include "templates/algorithms.h"
 
 // Return enum option info for ExpansionFunctionType
