+++ src/nodes/sq.h
@@ -7,13 +7,7 @@
 #include "math/function1D.h"
 #include "math/history.h"
 #include "math/windowFunction.h"
-#include "nodes/graph.h"
 #include "nodes/node.h"
-#include "nodes/parameter.h"
-
-// Forward Declarations
-class BraggModule;
-class PartialSet;
 
 class SQNode : public Node
 {
@@ -21,12 +15,15 @@ class SQNode : public Node
     SQNode(Graph *parentGraph);
     ~SQNode() override = default;
 
+    /*
+     * Definition
+     */
     public:
     std::string_view type() const override;
     std::string_view summary() const override;
 
     /*
-     * Definition
+     * Data
      */
     private:
     // Unweighted g(r)
@@ -37,8 +34,6 @@ class SQNode : public Node
     History<PartialSet> unweightedSQHistory_;
     // Number of historical partial sets to combine into final partials
     std::optional<Number> averagingLength_;
-    // Broadening function to apply to Bragg S(Q)
-    Function1DWrapper braggQBroadening_;
     // Broadening function to apply to S(Q)
     Function1DWrapper qBroadening_{Functions1D::Form::GaussianC2, {0.0, 0.02}};
     // Step size in Q for S(Q) calculation
@@ -52,23 +47,6 @@ class SQNode : public Node
     // Window function to use when Fourier-transforming reference S(Q) to g(r))
     WindowFunction::Form windowFunction_{WindowFunction::Form::None};
 
-    /*
-     * Functions
-     */
-    private:
-    // Calculate unweighted S(Q) from unweighted g(r)
-    bool calculateUnweightedSQ();
-
-    /*
-     * Processing
-     */
-    private:
-    // Run main processing
-    NodeConstants::ProcessResult process() override;
-
-    /*
-     * Getters
-     */
     public:
     // Returns the source configuration, belonging to the input GR node
     const Configuration *sourceConfiguration();
@@ -76,4 +54,11 @@ class SQNode : public Node
     const PartialSet &unweightedSQ() const;
     // Returns the unweighted GR
     const PartialSet &unweightedGR() const;
+
+    /*
+     * Processing
+     */
+    protected:
+    // Run main processing
+    NodeConstants::ProcessResult process() override;
 };
