+++ tests/nodes/bragg.cpp
@@ -7,7 +7,7 @@
 #include "math/rangedVector3.h"
 #include "nodes/gr.h"
 #include "nodes/iterableGraph.h"
-#include "nodes/sq/sq.h"
+#include "nodes/sq.h"
 #include "tests/graphData.h"
 #include "tests/testData.h"
 #include <gtest/gtest.h>
